Adicionando linhas no banco
